aarch64: address clippy lints and enable clippy CI (#938)
* Now that the aarch64 port is nearly complete,
  it's time to treat it as a first-class citizen by running
  aarch64 builds as a CI pass to ensure that other 
  changes don't unexpectedly break aarch64 builds.

+++ kernel/captain/src/lib.rs
@@ -221,7 +221,7 @@ pub fn init(
     // 2. Create the idle task for this CPU.
     spawn::create_idle_task()?;
     // 3. Cleanup bootstrap tasks, which handles this one and all other APs' bootstrap tasks.
-    spawn::cleanup_bootstrap_tasks(cpu_count as usize)?;
+    spawn::cleanup_bootstrap_tasks(cpu_count)?;
     // 4. "Finish" this bootstrap task, indicating it has exited and no longer needs to run.
     bootstrap_task.finish();
     // 5. Enable interrupts such that other tasks can be scheduled in.
